working version with front end + backend + weaviate running locally, divided ingestion code and go rest server code
