+++ services/ingestion/main.go
@@ -2,28 +2,24 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"log/slog"
-	"net/http"
 
-	"github.com/rs/cors"
 	"github.com/weaviate/weaviate-go-client/v4/weaviate"
-	"github.com/weaviate/weaviate-go-client/v4/weaviate/graphql"
-	"github.com/weaviate/weaviate/entities/models"
 
+	"mtguru/packages/config"
 	"mtguru/packages/custom_logger"
 )
 
-var activeConfig EnvironmentConfig
+var activeConfig config.EnvironmentConfig
 var client *weaviate.Client
 
 func init() {
 	custom_logger.CreateLogger()
-	activeConfig = CreateConfig()
+	activeConfig = config.CreateConfig()
 	client = createClient(activeConfig)
 }
 
-func createClient(conf EnvironmentConfig) *weaviate.Client {
+func createClient(conf config.EnvironmentConfig) *weaviate.Client {
 
 	cfg := weaviate.Config{
 		Host:   conf.WEAVIATE_URL,
@@ -49,102 +45,9 @@ func createClient(conf EnvironmentConfig) *weaviate.Client {
 
 }
 
-func searchDatabase(search_string string) *models.GraphQLResponse {
-
-	// search_string := "make my units fly"
-
-	ctx := context.Background()
-	response, err := client.GraphQL().Get().
-		WithClassName("Mtguru").
-		// WithFields is used to specify the fields you want to retrieve from the cards matched
-		WithFields(
-			graphql.Field{Name: "name"},
-			// graphql.Field{Name: "mana_cost"},
-			// graphql.Field{Name: "type_line"},
-			graphql.Field{Name: "oracle_text"},
-			// graphql.Field{Name: "power"},
-			// graphql.Field{Name: "toughness"},
-			// graphql.Field{Name: "loyalty"},
-			// graphql.Field{Name: "colors"},
-			graphql.Field{Name: "set_name"},
-			// graphql.Field{Name: "keywords"},
-			// graphql.Field{Name: "flavor_text"},
-			// graphql.Field{Name: "rarity"},
-			graphql.Field{Name: "scryfall_uri"},
-			graphql.Field{Name: "image_uris", Fields: []graphql.Field{
-				{Name: "normal"},
-				{Name: "large"},
-			}},
-			graphql.Field{Name: "_additional", Fields: []graphql.Field{
-				{Name: "distance"}}},
-		).
-		WithNearText(client.GraphQL().NearTextArgBuilder().
-			WithConcepts([]string{search_string})).
-		WithLimit(29).
-		Do(ctx)
-
-	if err != nil {
-		slog.Debug(err.Error())
-	}
-
-	slog.Info("Prompt:", "prompt", search_string)
-	slog.Info("Response:", "matches", response)
-
-	return response
-}
-
-func alive(w http.ResponseWriter, r *http.Request) {
-	slog.Info("I'm alive!")
-}
-
-type MTGuruSearchRequest struct {
-	Query   string            `json:"query"`
-	Filters map[string]string `json:"filters"`
-}
-
-func searchHandler(w http.ResponseWriter, r *http.Request) {
-
-	var requestBody MTGuruSearchRequest
-	err := json.NewDecoder(r.Body).Decode(&requestBody)
-	if err != nil {
-		slog.Debug("Error decoding request body", "error", err.Error())
-		http.Error(w, "Invalid request body", http.StatusBadRequest)
-		return
-	}
-
-	slog.Info("Received search request:", "query", requestBody.Query, "filters", requestBody.Filters)
-	results := searchDatabase(requestBody.Query)
-
-	responseJSON, err := json.Marshal(results)
-	if err != nil {
-		slog.Debug("Error marshalling response", "error", err.Error())
-		http.Error(w, "Internal server error", http.StatusInternalServerError)
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	w.Write(responseJSON)
-
-	// searchDatabase(requestBody.Query)
-}
-
-func initHandler() http.Handler {
-	mux := http.NewServeMux()
-
-	mux.HandleFunc("GET /api/health", alive)
-	mux.HandleFunc("POST /api/search", searchHandler)
-	return cors.Default().Handler(mux)
-
-}
-
 func main() {
-	// createIndex(client)
-	// populateIndex(client)
+	createIndex(client)
+	populateIndex(client)
 	// searchDatabase(client)
 	// updateCollection(client)
-	handler := initHandler()
-	slog.Info("Starting server on port 8888...")
-	http.ListenAndServe(":8888", handler)
-
 }
