+++ services/server/main.go
@@ -1,52 +1,144 @@
 package main
 
 import (
+	"context"
 	"encoding/json"
-	"fmt"
+	"log/slog"
+	"mtguru/packages/config"
+	"mtguru/packages/custom_logger"
 	"net/http"
-	"time"
 
 	"github.com/rs/cors"
+	"github.com/weaviate/weaviate-go-client/v4/weaviate"
+	"github.com/weaviate/weaviate-go-client/v4/weaviate/graphql"
+	"github.com/weaviate/weaviate/entities/models"
 )
 
-var listenerPort string = "8888"
+var activeConfig config.EnvironmentConfig
+var client *weaviate.Client
+
+func init() {
+	// init is called before main, so we can set up our logger and client here
+	custom_logger.CreateLogger()
+	activeConfig = config.CreateConfig()
+	client = createClient(activeConfig)
+}
+
+func createClient(conf config.EnvironmentConfig) *weaviate.Client {
+
+	cfg := weaviate.Config{
+		Host:   conf.WEAVIATE_URL,
+		Scheme: "http",
+		// AuthConfig: auth.ApiKey{Value: conf.WEAVIATE_API_KEY},
+		Headers: map[string]string{
+			"X-OpenAI-Api-Key": conf.OPEN_API_KEY,
+		},
+	}
+
+	client, err := weaviate.NewClient(cfg)
+	if err != nil {
+		slog.Debug(err.Error())
+	}
+
+	live, err := client.Misc().LiveChecker().Do(context.Background())
+	if err != nil {
+		slog.Debug(err.Error())
+	}
+
+	slog.Debug("%v", "is_remote_server_up", live)
+	return client
+
+}
+
+func searchDatabase(search_string string) *models.GraphQLResponse {
+
+	// search_string := "make my units fly"
+
+	ctx := context.Background()
+	response, err := client.GraphQL().Get().
+		WithClassName("Mtguru").
+		// WithFields is used to specify the fields you want to retrieve from the cards matched
+		WithFields(
+			graphql.Field{Name: "name"},
+			// graphql.Field{Name: "mana_cost"},
+			// graphql.Field{Name: "type_line"},
+			graphql.Field{Name: "oracle_text"},
+			// graphql.Field{Name: "power"},
+			// graphql.Field{Name: "toughness"},
+			// graphql.Field{Name: "loyalty"},
+			// graphql.Field{Name: "colors"},
+			graphql.Field{Name: "set_name"},
+			// graphql.Field{Name: "keywords"},
+			// graphql.Field{Name: "flavor_text"},
+			// graphql.Field{Name: "rarity"},
+			graphql.Field{Name: "scryfall_uri"},
+			graphql.Field{Name: "image_uris", Fields: []graphql.Field{
+				{Name: "normal"},
+				{Name: "large"},
+			}},
+			graphql.Field{Name: "_additional", Fields: []graphql.Field{
+				{Name: "distance"}}},
+		).
+		WithNearText(client.GraphQL().NearTextArgBuilder().
+			WithConcepts([]string{search_string})).
+		WithLimit(29).
+		Do(ctx)
+
+	if err != nil {
+		slog.Debug(err.Error())
+	}
+
+	slog.Info("Prompt:", "prompt", search_string)
+	slog.Info("Response:", "matches", response)
+
+	return response
+}
+
+type MTGuruSearchRequest struct {
+	Query   string            `json:"query"`
+	Filters map[string]string `json:"filters"`
+}
+
+func searchHandler(w http.ResponseWriter, r *http.Request) {
+
+	var requestBody MTGuruSearchRequest
+	err := json.NewDecoder(r.Body).Decode(&requestBody)
+	if err != nil {
+		slog.Debug("Error decoding request body", "error", err.Error())
+		http.Error(w, "Invalid request body", http.StatusBadRequest)
+		return
+	}
+
+	slog.Info("Received search request:", "query", requestBody.Query, "filters", requestBody.Filters)
+	results := searchDatabase(requestBody.Query)
+
+	responseJSON, err := json.Marshal(results)
+	if err != nil {
+		slog.Debug("Error marshalling response", "error", err.Error())
+		http.Error(w, "Internal server error", http.StatusInternalServerError)
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
+	w.Write(responseJSON)
+
+	// searchDatabase(requestBody.Query)
+}
 
 func initHandler() http.Handler {
 	mux := http.NewServeMux()
 
-	// Add route handlers
-	mux.HandleFunc("/time", getTime)
-
-	mux.HandleFunc("POST /clicks", postNumberOfClicks)
-
-	// cors.Default() setup the middleware with default options being
-	// all origins accepted with simple methods (GET, POST). See
-	// documentation below for more options.
+	// mux.HandleFunc("GET /api/health", alive)
+	mux.HandleFunc("POST /api/search", searchHandler)
 	return cors.Default().Handler(mux)
-}
 
-func getTime(w http.ResponseWriter, r *http.Request) {
-	fmt.Fprintf(w, time.Now().String())
-}
-
-func postNumberOfClicks(w http.ResponseWriter, r *http.Request) {
-	type Payload struct {
-		Clicks int `json:"clicks"`
-	}
-	var payload Payload
-	err := json.NewDecoder(r.Body).Decode(&payload)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-	fmt.Println("There have been", payload.Clicks, "clicks on the client!")
-	w.WriteHeader(http.StatusNoContent)
 }
 
 func main() {
 
 	handler := initHandler()
+	slog.Info("Starting server on port 8888...")
+	http.ListenAndServe(":8888", handler)
 
-	fmt.Println("Server listening at port", listenerPort)
-	http.ListenAndServe(":"+listenerPort, handler)
 }
